Added menu Class to menu.rb
